+++ Modules/Microsoft365DSC/DscResources/MSFT_SCFilePlanPropertyCitation/MSFT_SCFilePlanPropertyCitation.psm1
@@ -1,4 +1,4 @@
-using module ..\_Base\M365DSCResourceBase.psm1
+﻿using module ..\_Base\M365DSCResourceBase.psm1
 
 [DscResource()]
 class SCFilePlanPropertyCitation : M365DSCResourceBase
@@ -146,7 +146,11 @@ class SCFilePlanPropertyCitation : M365DSCResourceBase
             try
             {
                 $property = Get-FilePlanPropertyCitation | Where-Object -FilterScript { $_.Name -eq $this.Name }
-                if ($property.Mode.ToString() -ne 'PendingDeletion')
+                if ($null -eq $property)
+                {
+                    Write-Verbose -Message "Property $($this.Name) was not found."
+                }
+                elseif ("$($property.Mode)" -ne 'PendingDeletion')
                 {
                     Remove-FilePlanPropertyCitation -Identity $this.Name -Confirm:$false -ErrorAction Stop
                 }
@@ -157,7 +161,9 @@ class SCFilePlanPropertyCitation : M365DSCResourceBase
             }
             catch
             {
-                $this.LogError($_, $_)
+                $this.LogError($_, "Error removing the file plan property $($this.Name)")
+
+                throw
             }
         }
     }