+++ Modules/Microsoft365DSC/DscResources/MSFT_SCFilePlanPropertySubCategory/MSFT_SCFilePlanPropertySubCategory.psm1
@@ -1,4 +1,4 @@
-using module ..\_Base\M365DSCResourceBase.psm1
+﻿using module ..\_Base\M365DSCResourceBase.psm1
 
 [DscResource()]
 class SCFilePlanPropertySubCategory : M365DSCResourceBase
@@ -155,9 +155,13 @@ class SCFilePlanPropertySubCategory : M365DSCResourceBase
             {
                 $property = Get-FilePlanPropertySubCategory | Where-Object -FilterScript { $_.DisplayName -eq $this.Name -and `
                         $_.ParentId -eq $parentId }
-                if ($property.Mode.ToString() -ne 'PendingDeletion')
+                if ($null -eq $property)
                 {
-                    Remove-FilePlanPropertySubCategory -Identity $this.Name -Confirm:$false
+                    Write-Verbose -Message "Property $($this.Name) was not found."
+                }
+                elseif ("$($property.Mode)" -ne 'PendingDeletion')
+                {
+                    Remove-FilePlanPropertySubCategory -Identity $this.Name -Confirm:$false -ErrorAction Stop
                 }
                 else
                 {
@@ -166,7 +170,9 @@ class SCFilePlanPropertySubCategory : M365DSCResourceBase
             }
             catch
             {
-                $this.LogError($_, $_)
+                $this.LogError($_, "Error removing the file plan property $($this.Name)")
+
+                throw
             }
         }
     }